Refactoring and restructuring
