+++ tests/test_save.py
@@ -32,13 +32,14 @@ from pydicom.filereader import read_file_meta_info
 from pydicom.misc import is_dicom
 from pydicom.tag import ItemTag, SequenceDelimiterTag, Tag
 from pydicom.uid import UID, JPEGBaseline8Bit, generate_uid
+
 from wsidicom import WsiDicom
-from wsidicom.file.file import OffsetTableType
+from wsidicom.wsidicom_file import WsiDicomFile
+from wsidicom.file import OffsetTableType, WsiDicomFileWriter
 from wsidicom.geometry import Point, Size, SizeMm
 from wsidicom.image_data import ImageData, ImageOrigin
-from wsidicom.file import WsiDicomFile, WsiDicomFileWriter
 from wsidicom.uid import WSI_SOP_CLASS_UID
-from wsidicom.wsidicom import WsiDicomLevel
+from wsidicom.group.level import Level
 
 SLIDE_FOLDER = Path(os.environ.get("WSIDICOM_TESTDIR", "tests/testdata/slides"))
 
@@ -415,8 +416,8 @@ class WsiDicomFileSaveTests(unittest.TestCase):
         for _, wsi in self.test_folders.items():
             with TemporaryDirectory() as tempdir:
                 print(f"running test create child on {wsi.files}")
-                target = cast(WsiDicomLevel, wsi.levels[-2])
-                source = cast(WsiDicomLevel, wsi.levels[-3])
+                target = cast(Level, wsi.levels[-2])
+                source = cast(Level, wsi.levels[-3])
                 new_level = source.create_child(
                     2, Path(tempdir), generate_uid, 1, 100, "bot", 0
                 )
