+++ wsidicom/group/__init__.py
@@ -12,4 +12,5 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
-from wsidicom.image_data.image_data import ImageData, ImageOrigin
+from wsidicom.group.group import Group
+from wsidicom.group.level import Level
