+++ wsidicom/image_data.py
@@ -688,3 +688,61 @@ class ImageData(metaclass=ABCMeta):
             image.crop(box=cropped_tile_region.box_from_origin)
             tile_frame = self.encode(image)
         return tile_frame
+
+
+class PillowImageData(ImageData):
+    def __init__(self, image: PILImage):
+        self._image = image.convert("RGB")
+
+    @classmethod
+    def from_file(cls, file: Union[str, Path]) -> "PillowImageData":
+        image = Image.open(file)
+        return cls(image)
+
+    @property
+    def files(self) -> List[Path]:
+        filename = getattr(self._image, "filename", None)
+        if filename is None:
+            return []
+        return [filename]
+
+    @property
+    def transfer_syntax(self) -> UID:
+        return JPEGBaseline8Bit
+
+    @property
+    def image_size(self) -> Size:
+        return Size.from_tuple(self._image.size)
+
+    @property
+    def tile_size(self) -> Size:
+        return self.image_size
+
+    @property
+    def pixel_spacing(self) -> Optional[SizeMm]:
+        return None
+
+    @property
+    def samples_per_pixel(self) -> int:
+        return 3
+
+    @property
+    def photometric_interpretation(self) -> str:
+        return "YBR_FULL_422"
+
+    @property
+    def image_origin(self) -> ImageOrigin:
+        return super().image_origin
+
+    def _get_decoded_tile(self, tile_point: Point, z: float, path: str) -> PILImage:
+        if tile_point != Point(0, 0):
+            raise ValueError("Can only get Point(0, 0) from non-tiled image.")
+        return self._image
+
+    def _get_encoded_tile(self, tile: Point, z: float, path: str) -> bytes:
+        if tile != Point(0, 0):
+            raise ValueError("Can only get Point(0, 0) from non-tiled image.")
+        return self.encode(self._image)
+
+    def close(self):
+        self._image.close()
