+++ wsidicom/tile_index/sparse_tile_index.py
@@ -1,0 +1,182 @@
+#    Copyright 2021, 2022, 2023 SECTRA AB
+#
+#    Licensed under the Apache License, Version 2.0 (the "License");
+#    you may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
+
+from typing import Dict, List, Optional, Sequence, Set, Tuple
+
+import numpy as np
+from pydicom.dataset import Dataset
+
+from wsidicom.errors import WsiDicomNotFoundError
+from wsidicom.geometry import Point, Size
+from wsidicom.tile_index.tile_index import TileIndex
+
+
+class SparseTilePlane:
+    """Hold frame indices for the tiles in a sparse tiled file. Empty (sparse)
+    frames are represented by -1."""
+
+    def __init__(self, tiled_size: Size):
+        """Create a SparseTilePlane of specified size.
+
+        Parameters
+        ----------
+        tiled_size: Size
+            Size of the tiling
+        """
+        self._shape = tiled_size
+        self.plane = np.full(tiled_size.to_tuple(), -1, dtype=np.dtype(int))
+
+    def __repr__(self) -> str:
+        return f"{type(self).__name__}({self._shape})"
+
+    def __str__(self) -> str:
+        return self.pretty_str()
+
+    def __getitem__(self, position: Point) -> int:
+        """Get frame index from tile index at plane_position.
+
+        Parameters
+        ----------
+        plane_position: Point
+            Position in plane to get the frame index from
+
+        Returns
+        ----------
+        int
+            Frame index
+        """
+        frame_index = int(self.plane[position.x, position.y])
+        return frame_index
+
+    def __setitem__(self, position: Point, frame_index: int):
+        """Add frame index to tile index.
+
+        Parameters
+        ----------
+        plane_position: Point
+            Position in plane to add the frame index
+        frame_index: int
+            Frame index to add to the index
+        """
+        self.plane[position.x, position.y] = frame_index
+
+    def pretty_str(self, indent: int = 0, depth: Optional[int] = None) -> str:
+        return "Sparse tile plane"
+
+
+class SparseTileIndex(TileIndex):
+    """Index for mapping tile position to frame number for datasets containing
+    sparse tiles. Frame indices are retrieved from tile position, z, and path
+    by finding the corresponding matching SparseTilePlane (z and path) and
+    returning the frame index at tile position. If the tile is missing (due to
+    the sparseness), -1 is returned."""
+
+    def __init__(self, datasets: Sequence[Dataset]):
+        """Create sparse tile index for frames in datasets. Requires equal tile
+        size for all tile planes. Pixel data tiles are identified by the Per
+        Frame Functional Groups Sequence that contains tile colum, row, z,
+        path, and frame index. These are stored in a SparseTilePlane
+        (one plane for every combination of z and path).
+
+        Parameters
+        ----------
+        datasets: Sequence[Dataset]
+            List of datasets containing sparse tiled image data.
+        """
+        super().__init__(datasets)
+        self._planes = self._read_planes_from_datasets(datasets)
+        self._focal_planes = self._get_focal_planes()
+
+    @property
+    def focal_planes(self) -> List[float]:
+        return self._focal_planes
+
+    def __str__(self) -> str:
+        return self.pretty_str()
+
+    def pretty_str(self, indent: int = 0, depth: Optional[int] = None) -> str:
+        return (
+            f"Sparse tile index tile size: {self.tile_size}, "
+            f"plane size: {self.tiled_size}"
+        )
+
+    def get_frame_index(self, tile: Point, z: float, path: str) -> int:
+        """Return frame index for a Point tile, z coordinate, and optical
+        path.
+
+        Parameters
+        ----------
+        tile: Point
+            Tile xy to get.
+        z: float
+            Z coordinate to get.
+        path: str
+            ID of optical path to get.
+
+        Returns
+        ----------
+        int
+            Frame index.
+        """
+        try:
+            plane = self._planes[(z, path)]
+        except KeyError:
+            raise WsiDicomNotFoundError(f"Plane with z {z}, path {path}", str(self))
+        frame_index = plane[tile]
+        return frame_index
+
+    def _get_focal_planes(self) -> List[float]:
+        """Return list of focal planes defiend in planes.
+
+        Returns
+        ----------
+        List[float]
+            Focal planes, specified in um.
+        """
+        focal_planes: Set[float] = set()
+        for z, _ in self._planes.keys():
+            focal_planes.add(z)
+        return sorted(list(focal_planes))
+
+    def _read_planes_from_datasets(
+        self, datasets: Sequence[Dataset]
+    ) -> Dict[Tuple[float, str], SparseTilePlane]:
+        """Return SparseTilePlane from planes in datasets.
+
+        Parameters
+        ----------
+        datasets: Sequence[Dataset]
+           List of datasets to read planes from.
+
+        Returns
+        ----------
+        Dict[Tuple[float, str], SparseTilePlane]
+            Dict of planes with focal plane and optical identifier as key.
+        """
+        planes: Dict[Tuple[float, str], SparseTilePlane] = {}
+
+        for dataset in datasets:
+            frame_sequence = dataset.frame_sequence
+            for i, frame in enumerate(frame_sequence):
+                (tile, z) = self._read_frame_coordinates(frame)
+                identifier = dataset.read_optical_path_identifier(frame)
+
+                try:
+                    plane = planes[(z, identifier)]
+                except KeyError:
+                    plane = SparseTilePlane(self.tiled_size)
+                    planes[(z, identifier)] = plane
+                plane[tile] = i + dataset.frame_offset
+
+        return planes
