+++ wsidicom/wsidicom_web/__init__.py
@@ -1,4 +1,4 @@
-#    Copyright 2021 SECTRA AB
+#    Copyright 2023 SECTRA AB
 #
 #    Licensed under the Apache License, Version 2.0 (the "License");
 #    you may not use this file except in compliance with the License.
@@ -12,23 +12,5 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
-from wsidicom.image_data import ImageData
-from wsidicom.series import Labels, Levels, Overviews
-
-from .config import settings
-from .graphical_annotations import (
-    Annotation,
-    AnnotationGroup,
-    AnnotationInstance,
-    Measurement,
-    Point,
-    PointAnnotationGroup,
-    Polygon,
-    PolygonAnnotationGroup,
-    Polyline,
-    PolylineAnnotationGroup,
-)
-from .instance import WsiDataset, WsiInstance
-from .wsidicom import WsiDicom
-
-__version__ = "0.8.0"
+from wsidicom.wsidicom_web.wsidicom_web_client import WsiDicomWebClient
+from wsidicom.wsidicom_web.wsidicom_web_source import WsiDicomWebSource
