+++ wsidicom/wsidicom_web/wsidicom_web_client.py
@@ -1,5 +1,18 @@
+#    Copyright 2023 SECTRA AB
+#
+#    Licensed under the Apache License, Version 2.0 (the "License");
+#    you may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
+
 from enum import Enum
-from functools import cached_property
 from typing import List
 
 from dicomweb_client.api import DICOMwebClient
@@ -8,7 +21,6 @@ from pydicom import Dataset
 from pydicom.uid import UID
 from requests.auth import AuthBase
 
-from wsidicom.dataset import WsiDataset
 from wsidicom.uid import ANN_SOP_CLASS_UID, WSI_SOP_CLASS_UID
 
 
@@ -17,7 +29,7 @@ class DicomTags(Enum):
     SOP_INSTANCE_UID = "00080018"
 
 
-class DicomWebClient:
+class WsiDicomWebClient:
     def __init__(
         self, hostname: str, qido_prefix: str, wado_prefix: str, auth: AuthBase
     ):
@@ -63,31 +75,3 @@ class DicomWebClient:
             media_types=("image/jpeg",),
         )
         return frames[0]
-
-
-class WsiDicomWeb:
-    def __init__(
-        self, client: DicomWebClient, study_uid: UID, series_uid: UID, instance_uid: UID
-    ):
-        self._study_uid = study_uid
-        self._series_uid = series_uid
-        self._instance_uid = instance_uid
-        self._client = client
-
-    @cached_property
-    def dataset(self) -> WsiDataset:
-        dataset = self._client.get_instance(
-            self._study_uid, self._series_uid, self._instance_uid
-        )
-        self._dataset = WsiDataset(dataset)
-        return self._dataset
-
-    @property
-    def transfer_syntax(self) -> UID:
-        return UID(self.dataset["TransferSyntaxUID"].value)
-
-    def get_tile(self, tile: int) -> bytes:
-        # First frame for DICOM web is 1.
-        return self._client.get_frame(
-            self._study_uid, self._series_uid, self._instance_uid, tile + 1
-        )
